basics of string key mappings work

+++ lib/generic/mapping.h
@@ -1,13 +1,13 @@
 struct TYPE_STRUCT {
 	REFCOUNT_HEADER
-	int64_t len;
-	int64_t nentries;
+	size_t len;
+	size_t nentries;
 
 	// TODO: this is a bit hacky
 #define concat(a, b) a##b
 	struct concat(TYPE_STRUCT, _entry) {
 #undef concat
-		unsigned int hash;   // 0 if entry is unused
+		uint32_t hash;   // 0 if entry is unused
 		struct class_Str key;
 		ITEMTYPE value;
 	} *entries;
